+++ src/filter/time.rs
@@ -31,6 +31,10 @@ impl TimeFilter {
                             .and_local_timezone(Local)
                             .latest()
                     })
+                    .or_else(|| {
+                        let timestamp_secs = s.strip_prefix('@')?.parse().ok()?;
+                        DateTime::from_timestamp(timestamp_secs, 0).map(Into::into)
+                    })
                     .map(|dt| dt.into())
             })
     }
@@ -135,5 +139,32 @@ mod tests {
         assert!(!TimeFilter::after(&ref_time, t10s_before)
             .unwrap()
             .applies_to(&t1m_ago));
+
+        let ref_timestamp = 1707723412u64; // Mon Feb 12 07:36:52 UTC 2024
+        let ref_time = DateTime::parse_from_rfc3339("2024-02-12T07:36:52+00:00")
+            .unwrap()
+            .into();
+        let t1m_ago = ref_time - Duration::from_secs(60);
+        let t1s_later = ref_time + Duration::from_secs(1);
+        // Timestamp only supported via '@' prefix
+        assert!(TimeFilter::before(&ref_time, &ref_timestamp.to_string()).is_none());
+        assert!(
+            TimeFilter::before(&ref_time, &format!("@{}", ref_timestamp))
+                .unwrap()
+                .applies_to(&t1m_ago)
+        );
+        assert!(
+            !TimeFilter::before(&ref_time, &format!("@{}", ref_timestamp))
+                .unwrap()
+                .applies_to(&t1s_later)
+        );
+        assert!(
+            !TimeFilter::after(&ref_time, &format!("@{}", ref_timestamp))
+                .unwrap()
+                .applies_to(&t1m_ago)
+        );
+        assert!(TimeFilter::after(&ref_time, &format!("@{}", ref_timestamp))
+            .unwrap()
+            .applies_to(&t1s_later));
     }
 }