+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 - Add `dir` as an alias to `directory` when using `-t` \ `--type`, see #1460 and #1464 (@Ato2207).
+- Add support for @%s date format in time filters similar to GNU date (seconds since Unix epoch for --older/--newer), see #1493 (@nabellows)
 
 ## Bugfixes