+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ## Bugfixes
 
 - Respect NO_COLOR environment variable with `--list-details` option. (#1455)
+- Fix bug that would cause hidden files to be includeddespite gitignore rules
+  if search path is "." (#1461, BurntSushi/ripgrep#2711).
 
 
 ## Changes